Put RefToBaseToPtr_ioread in separate file
