+++ DataFormats/Common/interface/RefToBaseToPtr.h
@@ -27,30 +27,5 @@ namespace edm {
     }
     return Ptr<T>(ref.id(), ref.get(), ref.key());
   }
-
-  /*
-   * Do not use this outside an ioread rule in classes_def.xml
-   */
-  template <typename T>
-  Ptr<T> refToBaseToPtr_ioread(RefToBase<T> const& ref) {
-    if (ref.isNull()) {
-      return Ptr<T>();
-    }
-    if (ref.isTransient()) {
-      // This is a logic error, any ref being deserialized
-      // by construction would be persistent
-      return Ptr<T>();
-    } else {
-      EDProductGetter const* getter = ref.productGetter();
-      if (getter) {
-        return Ptr<T>(ref.id(), ref.key(), getter);
-      }
-    }
-    // If this is called in an iorule outside the framework, we cannot call
-    // ref.get(), but since outside the framework we can never fetch the ref,
-    // the Ptr will only be useful if accessed later from inside the framework.
-    // We can fill with a nullptr for now
-    return Ptr<T>(ref.id(), static_cast<const T*>(nullptr), ref.key());
-  }
 }  // namespace edm
 #endif
