+++ DataFormats/PatCandidates/src/classes_trigger.h
@@ -2,7 +2,7 @@
 #include "DataFormats/Common/interface/Association.h"
 #include "DataFormats/Common/interface/Wrapper.h"
 #include "DataFormats/Common/interface/PtrVector.h"
-#include "DataFormats/Common/interface/RefToBaseToPtr.h"
+#include "DataFormats/Common/interface/RefToBaseToPtr_ioread.h"
 
 #include "DataFormats/PatCandidates/interface/TriggerObjectStandAlone.h"
 #include "DataFormats/PatCandidates/interface/TriggerEvent.h"
